Restructure Dim.kt to avoid stack overflows
Previously when running Keigen on the roboRIO, the class hierarchy
created in Dim.kt would be enough to crash the roboRIO's JRE when trying
to load all the superclasses.

The structure is now flattened such that every type number is a subclass
of Num. This loses the order encoded in the class hierarchy, however for
hte purposes of matrix shape, that isn't necessary.